requirements.txt
Flask==2.2.5
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
mysql-connector-python==8.1.0
python-dotenv==1.0.0
